[Android] Fix issues setting Shell TabBar appearance. (#15734)
* Apply Android shell tab colors

* - add tests for Android
